requirements.txt
streamlit
openai
supabase
python-dotenv